requirements_dev.txt
aiohttp~=3.11
mypy>=1.8.0
pydantic>=2.9.2
pre-commit>=3.6.1
python-dotenv>=1.0.1
ruff~=0.15
mkdocs-material==9.7.1
mkdocstrings[python]==1.0.2
mashumaro>=3.13.1
orjson>=3.10.12